+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Remove all pool events listener at pool destroying.
+- Remove all worker node events listener at worker node destroying.
+
 ## [3.0.8] - 2023-11-25
 
 ### Fixed